feat: splitted selecting a command and execution
